restructure norms
