+++ transformer_engine/common/normalization/common.h
